ci: fix coverage reporting
It would be nice to test against the non-editable install version of the package but for the moment, I haven't read enough to fix it nicely. This solves the problem for now, but we should have something that confirms the built version of the package actually works (primarily to ensure the data files in ssssom folder are included).